move ownership logic into its own object
