+++ src/module/actor/SR5Actor.ts
@@ -47,6 +47,7 @@ import { MatrixICFlow } from './flows/MatrixICFlow';
 import { RollDataOptions } from '../item/Types';
 import { MatrixRebootFlow } from '../flows/MatrixRebootFlow';
 import { MatrixRules } from '@/module/rules/MatrixRules';
+import { ActorOwnershipFlow } from '@/module/actor/flows/ActorOwnershipFlow';
 
 /**
  * The general Shadowrun actor implementation, which currently handles all actor types.
@@ -2043,18 +2044,7 @@ export class SR5Actor<SubType extends Actor.ConfiguredSubType = Actor.Configured
      * @param uuid - uuid of the instance to check
      */
     async isOwnerOf(uuid: string): Promise<boolean> {
-        const device = await fromUuid(uuid) as any;
-        if (!device) return false;
-        if (device instanceof SR5Item) {
-            // check if we are the actorOwner of the item
-            return device.actorOwner === this;
-        } else if (device instanceof SR5Actor) {
-            if (device === this) return true;
-            // if it's a vehicle, check if we are the driver of it (this isn't perfect but will do for now)
-            const vehicle = device.asType('vehicle');
-            if (vehicle && vehicle.getVehicleDriver() === this) return true;
-        }
-        return false;
+        return ActorOwnershipFlow.isOwnerOf(this, uuid);
     }
 
     /**
